\name{enriched_goterms}
\alias{enriched_goterms}
\docType{data}
\title{
Sample Enriched Gene Ontologies (GO) Terms
}
\description{
This dataset highlights enriched gene ontologies (GO) terms identified by using ConsensusPathDB while performing overexpression analysis for a sample set of genes.
}
\usage{data("enriched_goterms")}
\references{
Kamburov, A., Stelzl, U., Lehrach, H. and Herwig, R., 2013. The ConsensusPathDB interaction database: 2013 update. Nucleic acids research, 41(D1), pp.D793-D800.
}
\examples{
data(enriched_goterms)
## maybe str(enriched_goterms) ; 
}
\keyword{datasets}
