\name{enriched_pathways}
\alias{enriched_pathways}
\docType{data}
\title{
Sample Enriched Pathways
}
\description{
This dataset highlights enriched pathways identified by using 'ConsensusPathDB' while performing overexpression analysis for a sample set of genes.
}
\usage{data("enriched_pathways")}
\references{
Kamburov, A., Stelzl, U., Lehrach, H. and Herwig, R., 2013. The ConsensusPathDB interaction database: 2013 update. Nucleic acids research, 41(D1), pp.D793-D800.
}
\examples{
data(enriched_pathways)
## maybe str(enriched_pathways) ; 
}
\keyword{datasets}
