\name{enriched_pc}
\alias{enriched_pc}
\docType{data}
\title{
Sample Enriched Protein Complexes
}
\description{
This dataset highlights protein complexes identified by using 'ConsensusPathDB' while performing overexpression analysis for a sample set of genes.
}
\usage{data("enriched_pc")}
\references{
Kamburov, A., Stelzl, U., Lehrach, H. and Herwig, R., 2013. The ConsensusPathDB interaction database: 2013 update. Nucleic acids research, 41(D1), pp.D793-D800.
}
\examples{
data(enriched_pc)
## maybe str(enriched_pc) ; 
}
\keyword{datasets}
